+++ CHANGELOG.md
@@ -29,6 +29,7 @@ All notable changes to this project will be documented in this file. This change
 - Log errors when initializing engine ([#787](https://github.com/mfikes/planck/issues/787))
 - Bundle missing macros namespaces ([#794](https://github.com/planck-repl/planck/issues/794))
 - ClojureScript version should not be displayed if there are inits ([#804](https://github.com/planck-repl/planck/issues/804))
+- Docstring for `planck.shell/sh` `:in` should indicate `string` ([#807](https://github.com/planck-repl/planck/issues/807))
 
 ## [2.17.0] - 2018-07-02
 ### Changed